requirements.txt
absl-py==1.0.0
altair==4.2.0
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
astunparse==1.6.3
attrs==21.4.0
Babel==2.9.1
backcall==0.2.0
black==22.1.0
bleach==4.1.0
blinker==1.5
bokeh==3.0.2
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.11
click==8.0.3
click-plugins==1.1.1
cligj==0.7.2
cmake==3.25.0
colorama==0.4.4
commonmark==0.9.1
comtypes==1.1.11
contourpy==1.0.6
cycler==0.11.0
Cython==0.29.28
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
dlib==19.24.0
easyocr==1.6.2
entrypoints==0.4
executing==0.8.2
face-recognition==1.3.0
face-recognition-models==0.3.0
filelock==3.8.0
Fiona==1.8.22
Flask==2.2.2
flatbuffers==2.0
fonttools==4.29.1
gast==0.5.3
gensim==3.8.0
geopandas==0.12.0
gevent==22.10.2
gitdb==4.0.10
GitPython==3.1.30
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googletrans==4.0.0rc1
greenlet==2.0.1
grpcio==1.43.0
gTTS==2.2.4
h11==0.9.0
h2==3.2.0
h5py==3.6.0
hpack==3.0.0
hstspreload==2021.12.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
imageio==2.15.0
imbalanced-learn==0.9.1
imblearn==0.0
importlib-metadata==6.0.0
imutils==0.5.4
ipykernel==6.9.0
ipython==8.0.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
json5==0.9.6
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyter-server==1.13.5
jupyterlab==3.3.1
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.3
jupyterlab-widgets==1.0.2
keras==2.8.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyboard==0.13.5
kiwisolver==1.3.2
lab==7.0
labelImg==1.8.6
libclang==13.0.0
llvmlite==0.39.1
lxml==4.8.0
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
missingno==0.5.1
mistune==0.8.4
MouseInfo==0.1.3
munch==2.5.0
mypy-extensions==0.4.3
mysql-connector-python==8.0.31
nbclassic==0.3.6
nbclient==0.5.10
nbconvert==6.4.1
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==2.6.3
ninja==1.11.1
nltk==3.7
notebook==6.4.8
notebook-shim==0.1.0
numba==0.56.4
numpy==1.24.1
oauthlib==3.2.0
opencv-contrib-python==4.7.0.68
opencv-python==4.5.5.62
opencv-python-headless==4.5.4.60
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.0
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
patsy==0.5.2
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==2.4.1
playsound==1.3.0
prometheus-client==0.13.1
prompt-toolkit==3.0.27
protobuf==3.19.4
psutil==5.9.4
pure-eval==0.2.2
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio @ file:///C:/Polyglot/PyAudio-0.2.11-cp310-cp310-win_amd64.whl
PyAutoGUI==0.9.53
pyclipper==1.3.0.post4
pycparser==2.21
pydeck==0.8.0
PyGetWindow==0.0.9
pyglet==1.5.23
Pygments==2.11.2
Pympler==1.0.1
PyMsgBox==1.0.9
pyparsing==3.0.7
pyperclip==1.8.2
pypiwin32==223
pyproj==3.4.0
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.1
PyRect==0.1.4
pyrsistent==0.18.1
PyScreeze==0.1.28
pytesseract==0.3.8
python-bidi==0.4.2
python-dateutil==2.8.2
pyttsx3==2.90
pytweening==1.0.4
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.2.0
pywin32==303
pywinpty==1.1.6
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.1
regex==2022.3.15
requests==2.27.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==13.1.0
rsa==4.8
scikit-image==0.19.1
scikit-learn==1.1.3
scipy==1.8.0
seaborn==0.11.2
semver==2.13.0
Send2Trash==1.8.0
Shapely==1.8.5.post1
simplejson==3.17.6
six==1.16.0
smart-open==6.2.0
smmap==5.0.0
sniffio==1.2.0
SpeechRecognition==3.8.1
stack-data==0.1.4
statsmodels==0.13.2
streamlit==1.17.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
terminado==0.13.1
testpath==0.5.0
tf-estimator-nightly==2.8.0.dev2021122109
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2022.2.9
tkcap==0.0.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.13.1
torchvision==0.14.1
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
txt2tags==3.7
typing_extensions==4.0.1
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.8
utils==1.0.1
validators==0.20.0
virtualenv==20.16.6
watchdog==2.2.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.1
Werkzeug==2.2.2
widgetsnbextension==3.5.2
wrapt==1.13.3
xyzservices==2022.9.0
zipp==3.11.0
zope.event==4.6
zope.interface==5.5.2
